requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
drf-spectacular>=0.22.1,<0.23
psycopg2>=2.8.6,<2.9
Pillow>=8.2.0,<8.3.0